make GCC happy

+++ src/Exception.cpp
@@ -330,7 +330,7 @@ void CJavascriptException::ThrowIf(v8::TryCatch& try_catch)
       {
         v8::String::Utf8Value s(v8::Handle<v8::String>::Cast(exc->Get(name)));
 
-        for (int i=0; i<_countof(SupportErrors); i++)
+        for (size_t i=0; i<_countof(SupportErrors); i++)
         {
           if (strnicmp(SupportErrors[i].name, *s, s.length()) == 0)
           {
